requirements.txt
nerfstudio==0.3.4
ultralytics==8.0.208
gdown==4.7.1
python-dotenv==1.0.0
supabase==2.0.3
python-box==6.0.2
ensure==1.0.2
-e .